requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pillow>=9.5.0
pytesseract>=0.3.10
spacy>=3.5.0
langchain>=0.0.200
langchain-openai>=0.0.1
openai>=0.27.0
pymongo>=4.3.3
motor>=3.1.1
numpy>=1.24.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
aiofiles>=23.1.0
jinja2>=3.1.2 